requirements.txt
selenium
pandas
python-dotenv
openai
schedule
streamlit
faker
faker
